+++ docs/campaign-love-is-the-law.md
@@ -1,0 +1,55 @@
+# Love Is The Law — Supreme Court Campaign
+
+A campaign for the Supreme Court game focused on the legal history of gay rights in America, inspired by the Minneapolis band The Suburbs.
+
+## Campaign Overview
+
+- **Title:** Love Is The Law
+- **President:** Barack Obama
+- **Chief Justice:** Antonin Scalia
+- **Theme:** Argue for or against gay rights as the Supreme Court hears the most pivotal cases in LGBTQ+ legal history.
+
+## Cases (Chronological Order)
+
+1. **Bowers v. Hardwick (1986)** — To be added
+2. **Romer v. Evans (1996)** — To be added
+3. **Lawrence v. Texas (2003)** — To be added
+4. **Obergefell v. Hodges (2015)**
+5. **Masterpiece Cakeshop v. Colorado (2018)**
+
+## Bench (9 Justices, with stances)
+
+- Antonin Scalia (Against)
+- Anthony Kennedy (For, to be added)
+- Ruth Bader Ginsburg (For, to be added)
+- Sonia Sotomayor (For, to be added)
+- Samuel Alito (Against, to be added)
+- Clarence Thomas (Against)
+- [Right-leaning/historical justice]
+- [Right-leaning/historical justice]
+- [Right-leaning/historical justice]
+
+_If not enough justices have explicit "GayRights" stances, fill with right-leaning/historical justices._
+
+## Required Data Changes
+
+- **Add/annotate "GayRights" stances** for:
+  - Anthony Kennedy (For)
+  - Ruth Bader Ginsburg (For)
+  - Sonia Sotomayor (For)
+  - Samuel Alito (Against)
+- **Add three new cases**:
+  - Bowers v. Hardwick (1986)
+  - Romer v. Evans (1996)
+  - Lawrence v. Texas (2003)
+- **Campaign config**: Add to `campaignSetups` in `_campaigns.ts` with the above details.
+
+## Notes
+
+- This campaign is specifically about the legal journey of gay rights, not general equality.
+- The bench should reflect a mix of strong opinions on gay rights, with a right-leaning tilt if needed for balance.
+- All changes should be reviewed before merging.
+
+---
+
+_Drafted for future implementation. See this file for requirements and context when creating the campaign or related issue._
