+++ src/views/court/ts/_types.ts
@@ -34,44 +34,6 @@ export type StanceType =
   | "WomensRights" // Legal equality and protections specific to women
   | "WorkersRights"; // Labor protections, workplace safety, and employee rights
 
-/** All valid StanceType values as a runtime array for iteration and validation. */
-export const ALL_STANCE_TYPES: StanceType[] = [
-  "AccusedRights",
-  "AgencyDeference",
-  "AnimalRights",
-  "CampaignFinance",
-  "ChristianHegemony",
-  "CivilLiberties",
-  "CourtAuthority",
-  "CorporatePower",
-  "Drugs",
-  "EconomicRegulation",
-  "EnvironmentalProtection",
-  "ExecutivePower",
-  "FairElections",
-  "FederalPower",
-  "FreeSpeech",
-  "GayRights",
-  "Guns",
-  "Immigration",
-  "JudicialActivism",
-  "NativeSovereignty",
-  "Privacy",
-  "PropertyRights",
-  "PolicePower",
-  "RacialEquity",
-  "ReligiousLiberty",
-  "ReproductiveRights",
-  "RichPeople",
-  "SocialOrder",
-  "StatesRights",
-  "Taxation",
-  "VotingRights",
-  "WarPowers",
-  "WomensRights",
-  "WorkersRights",
-];
-
 export type StanceOpinion = "For" | "Against" | "Neutral";
 
 export interface Justice {
