+++ tests/court-stances.spec.ts
@@ -8,7 +8,6 @@
  */
 
 import { describe, it, expect } from "vitest";
-import { ALL_STANCE_TYPES } from "../src/views/court/ts/_types";
 import type { StanceType } from "../src/views/court/ts/_types";
 import {
   justiceCurrent,
@@ -20,22 +19,41 @@ import {
 } from "../src/views/court/ts/_justices";
 import { presidents } from "../src/views/court/ts/_presidents";
 import { cases as allCases, casesHistorical, casesFictional } from "../src/views/court/ts/_cases";
+import Table from "cli-table3";
+import colors from "colors";
 
 const casesArray = [...allCases, ...casesHistorical, ...casesFictional];
+
+// Dynamically collect all unique stance topics from justices, presidents, and cases
+const stanceSet = new Set<string>();
+const addStances = (arr: any[], key: string = "stances") => {
+  arr.forEach((item) => {
+    (item[key] ?? []).forEach(({ topic }: { topic: string }) => {
+      stanceSet.add(topic);
+    });
+  });
+};
+addStances([...justiceCurrent, ...justiceHistorical, ...justiceFictional, ...justiceCelebrity, ...justiceWarrenExtra, ...justiceLochnerExtra]);
+addStances(presidents);
+casesArray.forEach((c) => {
+  Object.keys(c.prosecution.stances ?? {}).forEach((topic) => stanceSet.add(topic));
+  Object.keys(c.defendant.stances ?? {}).forEach((topic) => stanceSet.add(topic));
+});
+const ALL_STANCE_TOPICS: string[] = Array.from(stanceSet).sort();
 
 // ─── Build usage counts ──────────────────────────────────────────────────────
 
 const counts: Record<
-  StanceType,
+  string,
   {
     justices: number;
     presidents: number;
     cases: number;
     total: number;
   }
-> = {} as never;
+> = {};
 
-ALL_STANCE_TYPES.forEach((t) => {
+ALL_STANCE_TOPICS.forEach((t) => {
   counts[t] = { justices: 0, presidents: 0, cases: 0, total: 0 };
 });
 
@@ -83,55 +101,121 @@ presidents.forEach((p) => {
 
 // ─── Sorted arrays ───────────────────────────────────────────────────────────
 
-const sorted = ALL_STANCE_TYPES.slice().sort((a, b) => counts[b].total - counts[a].total);
+const sorted = ALL_STANCE_TOPICS.slice().sort((a, b) => counts[b].total - counts[a].total);
 const top5 = sorted.slice(0, 5);
 const bottom5 = sorted.slice(-5).reverse();
 const unused = sorted.filter((t) => counts[t].total === 0);
 
-// ─── Console output ──────────────────────────────────────────────────────────
+// ─── CLI-TABLE3 + COLORS OUTPUT ─────────────────────────────────────────────
 
-console.log("\n══════════════════════════════════════════");
-console.log("  STANCE USAGE REPORT — Court Card Game  ");
-console.log("══════════════════════════════════════════\n");
+console.log("\n" + colors.cyan.bold("================================================================"));
+console.log(colors.cyan.bold("  STANCE USAGE REPORT -- Court Card Game  "));
+console.log(colors.cyan.bold("================================================================\n"));
 
-console.log(`Total stances tracked: ${ALL_STANCE_TYPES.length}`);
-console.log(`Total justice stances: ${allJustices.reduce((n, j) => n + (j.stances?.length ?? 0), 0)}`);
-console.log(`Total president stances: ${presidents.reduce((n, p) => n + (p.stances?.length ?? 0), 0)}`);
-const caseStanceCount = casesArray.reduce((n, c) => n + Object.keys(c.prosecution.stances ?? {}).length + Object.keys(c.defendant.stances ?? {}).length, 0);
-console.log(`Total case stances: ${caseStanceCount}\n`);
+console.log(colors.white("Total stances tracked: ") + colors.yellow(String(ALL_STANCE_TOPICS.length)));
+console.log(
+  colors.white("Total justice stances: ") +
+    colors.yellow(
+      String(
+        allJustices.reduce(function (n, j) {
+          return n + (j.stances ? j.stances.length : 0);
+        }, 0),
+      ),
+    ),
+);
+console.log(
+  colors.white("Total president stances: ") +
+    colors.yellow(
+      String(
+        presidents.reduce(function (n, p) {
+          return n + (p.stances ? p.stances.length : 0);
+        }, 0),
+      ),
+    ),
+);
+const caseStanceCount = casesArray.reduce(function (n, c) {
+  return n + Object.keys(c.prosecution.stances || {}).length + Object.keys(c.defendant.stances || {}).length;
+}, 0);
+console.log(colors.white("Total case stances: ") + colors.yellow(String(caseStanceCount)) + "\n");
 
-console.log("🔝 Top 5 Most Used Stances:");
+// Top 5 Table
+const topTable = new Table({
+  head: [
+    colors.green.bold("#"),
+    colors.green.bold("Stance Type"),
+    colors.green.bold("Total"),
+    colors.green.bold("Justices"),
+    colors.green.bold("Presidents"),
+    colors.green.bold("Cases"),
+  ],
+  colWidths: [4, 26, 8, 10, 12, 8],
+});
 top5.forEach((t, i) => {
   const { justices, presidents: pres, cases: cs, total } = counts[t];
-  console.log(`  ${i + 1}. ${t.padEnd(25)} total=${total}  (justices=${justices}, presidents=${pres}, cases=${cs})`);
+  topTable.push([
+    colors.green((i + 1).toString()),
+    colors.bold(t),
+    colors.yellow(total.toString()),
+    colors.cyan(justices.toString()),
+    colors.magenta(pres.toString()),
+    colors.blue(cs.toString()),
+  ]);
 });
+console.log(colors.green.bold("🔝 Top 5 Most Used Stances:"));
+console.log(topTable.toString());
 
-console.log("\n🔻 Bottom 5 Least Used Stances:");
+// Bottom 5 Table
+const bottomTable = new Table({
+  head: [
+    colors.red.bold("#"),
+    colors.red.bold("Stance Type"),
+    colors.red.bold("Total"),
+    colors.red.bold("Justices"),
+    colors.red.bold("Presidents"),
+    colors.red.bold("Cases"),
+  ],
+  colWidths: [4, 26, 8, 10, 12, 8],
+});
 bottom5.forEach((t, i) => {
   const { justices, presidents: pres, cases: cs, total } = counts[t];
-  console.log(`  ${i + 1}. ${t.padEnd(25)} total=${total}  (justices=${justices}, presidents=${pres}, cases=${cs})`);
+  bottomTable.push([
+    colors.red((i + 1).toString()),
+    colors.bold(t),
+    colors.yellow(total.toString()),
+    colors.cyan(justices.toString()),
+    colors.magenta(pres.toString()),
+    colors.blue(cs.toString()),
+  ]);
 });
+console.log("\n" + colors.red.bold("🔻 Bottom 5 Least Used Stances:"));
+console.log(bottomTable.toString());
 
+// Unused stances
 if (unused.length > 0) {
-  console.log(`\n⚠️  UNUSED STANCES (${unused.length}) — these may be candidates for removal or need more coverage:`);
-  unused.forEach((t) => console.log(`  • ${t}`));
+  console.log("\n" + colors.bgYellow.black("UNUSED STANCES (" + unused.length + ") — these may be candidates for removal or need more coverage:"));
+  const unusedTable = new Table({
+    head: [colors.yellow.bold("Unused Stance Type")],
+    colWidths: [32],
+  });
+  unused.forEach((t) => unusedTable.push([colors.yellow(t)]));
+  console.log(unusedTable.toString());
 } else {
-  console.log("\n✅ All stance types are used at least once.");
+  console.log("\n" + colors.green.bold("✅ All stance types are used at least once."));
 }
-console.log("\n══════════════════════════════════════════\n");
+console.log("\n" + colors.cyan.bold("════════════════════════════════════════════════════════════════\n"));
 
 // ─── Tests ───────────────────────────────────────────────────────────────────
 
 describe("Stance Usage", () => {
   it("has at least 30 stance types defined", () => {
-    expect(ALL_STANCE_TYPES.length).toBeGreaterThanOrEqual(30);
+    expect(ALL_STANCE_TOPICS.length).toBeGreaterThanOrEqual(30);
   });
 
   it("every justice stance topic is a valid StanceType", () => {
     const invalidJusticeStances: string[] = [];
     allJustices.forEach((j) => {
       (j.stances ?? []).forEach(({ topic }) => {
-        if (!ALL_STANCE_TYPES.includes(topic as StanceType)) {
+        if (!ALL_STANCE_TOPICS.includes(topic)) {
           invalidJusticeStances.push(`Justice ${j.name}: "${topic}"`);
         }
       });
@@ -146,7 +230,7 @@ describe("Stance Usage", () => {
     const invalidPresidentStances: string[] = [];
     presidents.forEach((p) => {
       (p.stances ?? []).forEach(({ topic }) => {
-        if (!ALL_STANCE_TYPES.includes(topic as StanceType)) {
+        if (!ALL_STANCE_TOPICS.includes(topic)) {
           invalidPresidentStances.push(`President ${p.name}: "${topic}"`);
         }
       });
@@ -161,7 +245,7 @@ describe("Stance Usage", () => {
     const invalidCaseStances: string[] = [];
     casesArray.forEach((c) => {
       [...Object.keys(c.prosecution.stances ?? {}), ...Object.keys(c.defendant.stances ?? {})].forEach((topic) => {
-        if (!ALL_STANCE_TYPES.includes(topic as StanceType)) {
+        if (!ALL_STANCE_TOPICS.includes(topic)) {
           invalidCaseStances.push(`Case "${c.name}": "${topic}"`);
         }
       });
